fix RefillPickingIfEmptyJobTest & BasicModulesTest
