+++ tests/Modules/AutoStatusPicking/src/Jobs/RefillPickingIfEmptyJobTest.php
@@ -2,13 +2,33 @@
 
 namespace Tests\Modules\AutoStatusPicking\src\Jobs;
 
+use App\Models\Order;
+use App\Models\OrderProduct;
+use App\Modules\AutoStatusPicking\src\Jobs\RefillPickingIfEmptyJob;
+use Carbon\Carbon;
 use Tests\TestCase;
 
 class RefillPickingIfEmptyJobTest extends TestCase
 {
-    /** @test */
-    public function test_job()
+    /**
+     * A basic feature test example.
+     */
+    public function testExample(): void
     {
-        $this->markTestIncomplete('This test has not been implemented yet');
+        /** @var Order $order */
+        $order = Order::factory()->create([
+            'order_placed_at' => Carbon::now()->subDays(100),
+            'status_code' => 'processing',
+        ]);
+
+        OrderProduct::factory(2)->create([
+            'order_id' => $order->id,
+        ]);
+
+        $order->update(['status_code' => 'paid']);
+
+        RefillPickingIfEmptyJob::dispatchSync();
+
+        $this->assertDatabaseHas('orders', ['status_code' => 'picking']);
     }
 }
